+++ __init__.py
@@ -4,7 +4,7 @@ bl_info = {
     "name": "Storytools - Storyboard Tools",
     "description": "Set of tools for Storyboarding",
     "author": "Samuel Bernou",
-    "version": (1, 17, 0),
+    "version": (1, 17, 2),
     "blender": (4, 0, 0),
     "location": "View3D",
     "warning": "",
@@ -26,7 +26,7 @@ from . import gizmos_camera
 from . import gizmo_toolbar
 from . import map
 from . import handles
-from . import panels
+from . import ui
 from . import keymaps
 from . import gizmo_toolpreset_bar
 from .fn import get_addon_prefs
@@ -44,7 +44,7 @@ modules = (
     gizmos_camera,
     gizmo_toolbar,
     # map,
-    panels,
+    ui,
     keymaps,
     gizmo_toolpreset_bar,
 )